osm2mrag/requirements.txt
shapely
psycopg2-binary
python-dotenv
